requirements.txt
pysimplegui
pillow
PyPDF3
